project-strava-api/requirements.txt
# FastAPI et dépendances
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Authentification et sécurité
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# Validation
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Génération de documents
reportlab==4.0.7
matplotlib==3.8.2
pillow==10.1.0

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-mock==3.12.0

# Utilitaires
click==8.1.7
rich==13.7.0